layer-content 컴포넌트 제거
